+++ src/chimeric/mod.rs
@@ -16,7 +16,9 @@ mod output;
 mod score;
 mod segment;
 
-pub use detect::{ChimericDetector, detect_chimeric_old, detect_inter_mate_chimeric};
+pub use detect::{
+    ChimericDetector, detect_chimeric_mult, detect_chimeric_old, detect_inter_mate_chimeric,
+};
 pub use output::{ChimericJunctionWriter, build_within_bam_records};
 pub use segment::{ChimericAlignment, ChimericSegment};